+++ BthPS3/L2CAP.h
@@ -69,11 +69,6 @@ EVT_WDF_REQUEST_COMPLETION_ROUTINE L2CAP_PS3_DenyRemoteConnectCompleted;
 EVT_WDF_REQUEST_COMPLETION_ROUTINE L2CAP_PS3_ControlConnectResponseCompleted;
 EVT_WDF_REQUEST_COMPLETION_ROUTINE L2CAP_PS3_InterruptConnectResponseCompleted;
 
-_IRQL_requires_max_(DISPATCH_LEVEL)
-NTSTATUS
-L2CAP_PS3_ConnectionStateConnected(
-    PBTHPS3_CLIENT_CONNECTION ClientConnection
-);
 
 _IRQL_requires_max_(DISPATCH_LEVEL)
 NTSTATUS
